Got feature parity on refactor with master
