+++ machine_usage/templates/machine_usage/forms/machine_usage.html
@@ -6,7 +6,7 @@
 {% endblock page_title %}
 
 {% block custom_styles %}
-    <link rel="stylesheet" type="text/css" href="{% static 'machine_usage/css/forms/machine_usage.css' %}">
+    <link rel="stylesheet" type="text/css" href="{% static 'forge/css/forms/machine_usage.css' %}">
 {% endblock custom_styles %}
     
 {% block header %}
