--- machine_usage/templates/machine_usage/login.html
@@ -1,27 +1,0 @@
-{% extends "machine_usage/generic.html" %}
-{% load static %}
-
-{% block page_title %}
-<title>Log In</title>
-{% endblock page_title %}
-
-{% block custom_styles %}
-	<link rel="stylesheet" type="text/css" href="{% static 'machine_usage/css/login.css' %}">
-{% endblock custom_styles %}
-	
-{% block main %}
-	<main class="bg_shade">
-		<div id="login_form_container" class="center_children fill_parent_horiz fill_screen_vert">
-			<div id="login_form">
-				<h1>Sign In</h1>
-				<form method="POST">
-					{% csrf_token %}
-					<input type="text" name="rcsid" placeholder="RCS ID" required autofocus /><br />
-					<input type="password" name="password" placeholder="Password" required/><br />
-					<button id="sign_in_button">Sign In</button>
-					<button id="create_account_button" onclick="location.href='forms/create_user';return false;">Create Account</button>
-				</form>
-			</div>
-		</div>
-	</main>
-{% endblock main %}
