--- machine_usage/templates/machine_usage/news.html
@@ -1,20 +1,0 @@
-{% extends "machine_usage/generic.html" %}
-{% load static %}
-
-{% block page_title %}
-<title>Forge News</title>
-{% endblock page_title %}
-
-{% block custom_styles %}
-	<link href="https://assets.juicer.io/embed.css" media="all" rel="stylesheet" type="text/css" />
-{% endblock custom_styles %}
-
-{% block main %}
-<main>
-	<ul class="juicer-feed" data-feed-id="653114304851265"><h1 class="referral"><a href="https://www.juicer.io">Powered by Juicer.io</a></h1></ul>
-</main>
-{% endblock main %}
-
-{% block scripts %}
-<script src="https://assets.juicer.io/embed.js" type="text/javascript"></script>
-{% endblock scripts %}
