+++ myforge/templates/myforge/myforge.html
@@ -1,4 +1,4 @@
-{% extends "machine_usage/generic.html" %}
+{% extends "forge/generic.html" %}
 {% load static %}
 {% load in_group %}
 
@@ -7,7 +7,7 @@
 {% endblock page_title %}
 
 {% block custom_styles %}
-	<link rel="stylesheet" type="text/css" href="{% static 'machine_usage/css/myforge.css' %}">
+	<link rel="stylesheet" type="text/css" href="{% static 'forge/css/myforge.css' %}">
 {% endblock custom_styles %}
 
 {% block pre_main %}
