+++ user_management/templates/user_management/begin_semester.html
@@ -1,4 +1,4 @@
-{% extends "machine_usage/generic.html" %}
+{% extends "forge/generic.html" %}
 {% load static %}
 
 {% block page_title %}
@@ -6,7 +6,7 @@
 {% endblock page_title %}
 
 {% block custom_styles %}
-	<link rel="stylesheet" type="text/css" href="{% static 'machine_usage/css/begin_semester.css' %}">
+	<link rel="stylesheet" type="text/css" href="{% static 'forge/css/begin_semester.css' %}">
 {% endblock custom_styles %}
 	
 {% block main %}
